Add a separate test suite to validate compatibility w/ GE plugin
