+++ src/enterpriseCompatibilityTest/java/com/gradle/ccud/adapters/enterprise/proxies/BuildScanCaptureSettingsProxyTest.java
@@ -1,7 +1,6 @@
-package com.gradle.ccud.adapters.enterprise;
+package com.gradle.ccud.adapters.enterprise.proxies;
 
 import com.gradle.ccud.adapters.reflection.ProxyFactory;
-import com.gradle.ccud.adapters.enterprise.proxies.BuildScanCaptureSettingsProxy;
 import com.gradle.scan.plugin.BuildScanCaptureSettings;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
