+++ src/enterpriseCompatibilityTest/java/com/gradle/ccud/adapters/enterprise/proxies/GradleEnterpriseExtensionProxyTest.java
@@ -1,7 +1,6 @@
-package com.gradle.ccud.adapters.enterprise;
+package com.gradle.ccud.adapters.enterprise.proxies;
 
 import com.gradle.ccud.adapters.reflection.ProxyFactory;
-import com.gradle.ccud.adapters.enterprise.proxies.GradleEnterpriseExtensionProxy;
 import com.gradle.enterprise.gradleplugin.GradleEnterpriseExtension;
 import com.gradle.scan.plugin.BuildScanExtension;
 import org.junit.jupiter.api.BeforeEach;
